move abstract and interface
