+++ core-bundle/src/DataContainer/BuilderTemplate/DataContainerBuilderTemplateInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Contao\CoreBundle\DataContainer;
+namespace Contao\CoreBundle\DataContainer\BuilderTemplate;
 
 interface DataContainerBuilderTemplateInterface
 {
