+++ core-bundle/src/DataContainer/BuilderTemplate/TitleBuilderTemplate.php
@@ -2,7 +2,6 @@
 
 namespace Contao\CoreBundle\DataContainer\BuilderTemplate;
 
-use Contao\CoreBundle\DataContainer\AbstractDataContainerBuilderTemplate;
 use Contao\DataContainer;
 
 /**
