+++ core-bundle/src/DataContainer/DataContainerBuilderFactory.php
@@ -2,6 +2,7 @@
 
 namespace Contao\CoreBundle\DataContainer;
 
+use Contao\CoreBundle\DataContainer\BuilderTemplate\DataContainerBuilderTemplateInterface;
 use Contao\System;
 
 class DataContainerBuilderFactory implements DataContainerBuilderFactoryInterface
